PR edits from review comments

+++ datahub-graphql-core/src/main/java/com/linkedin/datahub/graphql/resolvers/subscription/DeleteSubscriptionResolver.java
@@ -41,6 +41,14 @@ public class DeleteSubscriptionResolver implements DataFetcher<CompletableFuture
           throw new RuntimeException(
               String.format("Unauthorized to delete subscription for group %s", actorUrn));
         }
+        if (actorUrn.getEntityType().equals(CORP_USER_ENTITY_NAME) && !actorUrn.toString().equals(context.getActorUrn())) {
+          throw new RuntimeException(
+              String.format(
+                  "Unauthorized to delete personal subscription for actor user is not logged in as. User: %s, Subscription actor: %S",
+                  context.getActorUrn(),
+                  actorUrn
+              ));
+        }
 
         _entityClient.deleteEntity(subscriptionUrn, authentication);
